requirements.txt
aiogram==3.4.1
apscheduler
pillow

